refactor(frontend): move the flag panel's state into a hook
FlagListPanel combined permission checks, loading, fetching, creation,
resolution, notifications and rendering in one component. Move the state into
usePatientFlags so the panel is a thin projection onto the presentational list,
matching the waitlist and allergy panels.

Behaviour is unchanged; the panel's own tests pass untouched (312 tests across
the companion-history suites). tsc 0, lint clean.
